requirement.txt
flask
google-cloud-vision
pymongo